Added the Apply method
This method applies a given function to all records in the database. The
function must match the ApplyFunc type:

    func (key string, value interface{}, args ...interface{})

The function is called with the key, the value and extra arguments
passed to Apply. For instance, if you have a database with the following
records:

    x: 1
    y: 2

and you call the Apply this way:

    func ToApply(key string, value interface{}, args ...interface{}) {
        fmt.Printf("%s: %s\nFirst arg: %s\n\n", key, value.(string), args[0].(string))
    }

    db.Apply(ToApply, "gopher")

then you will get the following output:

    x: 1
    First arg: gopher

    y: 2
    First arg: gopher