Rewrite the KT bindings in Go.
The gokabinet bindings were an often point of pain and would randomly
fail in tests with memory errors. On top of that, they made us
require cgo.

I've rewritten the interfacing code in go using the spec available
at http://fallabs.com/kyototycoon/spex.html

I see the main benefit of this change being that we can more easily
change the behavior of the library. Some of the design decisions in
gokabinet were downright archaic and we finally have a chance to get
rid of stuff like errors on success.