kt: Retry request if remote end performs graceful restart
A HTTP1 server closes all idle connections when it performs a graceful shutdown. This is a problem for things like http.Transport, which maintain a pool of (possibly idle) connections, since an attempt to re-use such a closed connection will return io.EOF to the caller. This makes it impossible to implement zero-downtime restarts which do not drop requests.

This change treats io.EOF as a signal that a zero-downtime restart is in progress. It closes all other idle connections, and retries the request which yielded io.EOF once.