add thread-safe tockenbucket touch() function
Making the touch() thread-safe allows removing external mutex(es) that
protect calling the touch.

Unarguably the change also makes the data racing more, but that is fine.
With low frequency updates the tocket bucket will not trip over, and with
high frequency updates one should not be able to observe off-by-one type of
return value distortion when values are stored few jiffies apart.

Signed-off-by: Sami Kerola <kerolasa@cloudflare.com>